Interazione con i nodi per conoscere i dettagli dei flussi e divisione dei file in moduli
